Remove dom interaction logic into a file
