+++ src/utils/layoutTransition/index.ts
@@ -1,3 +1,4 @@
+import fixElementPosition from './fixElementPosition';
 import {ifMarkedThenMeasure, markAndMeasure} from './markAndMeasure';
 
-export {ifMarkedThenMeasure, markAndMeasure};
+export {ifMarkedThenMeasure, markAndMeasure, fixElementPosition};
